Refactor loading page #74
