+++ src/pages/loading.page.tsx
@@ -1,7 +1,7 @@
 import React, { ReactElement } from "react"
 
 import { BaseContainer } from "../components/ui/containers"
-import { Logo } from "../components/logo/logo"
+import { Logo } from "../components/header/logo"
 import { LoadingMessage } from "../components/loading/loading-message"
 
 export const LoadingPage = (): ReactElement => (
